requirements.txt
Flask==3.0.0
deep-translator==1.11.4
python-docx==1.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
mammoth==1.7.0